requirements.txt
fastapi==0.115.8
sqlalchemy==2.0.38
argon2-cffi==23.1.0
cachetools==5.5.2